move LanguageName & LocalizedName to ui/common package
Relocates shared language name composables from `ui/languages` to
`ui/common`, updating package declarations and imports across all
consumers.
