+++ app/src/main/kotlin/org/cru/godtools/ui/common/LanguageName+Previews.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.ui.languages
+package org.cru.godtools.ui.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.HorizontalDivider
