+++ app/src/test/snapshots/images/org.cru.godtools.ui.common_LanguageNamePaparazziTest_LanguageNames()[NIGHT].png
