+++ app/src/testDebug/kotlin/org/cru/godtools/ui/common/LanguageNamePaparazziTest.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.ui.languages
+package org.cru.godtools.ui.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
